# .openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CWS Next App API",
    "version": "0.1.0",
    "description": "Automated API documentation for the CWS Next App. This document is generated from Zod schemas and route handler metadata.",
    "contact": {
      "name": "API Support",
      "email": "api@crossweavesourcing.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
...
  "paths": {
...
    "/api/auth/webauthn/login-options": {
      "post": {
        "operationId": "webauthnLoginOptions",
        "summary": "Get WebAuthn login options",
        "description": "Generates passwordless passkey authentication options for this enrolled device.",
        "tags": [
          "Authentication"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
                    "description": "Account email for email-first passkey login",
                    "example": "admin@example.com"
                  }
                },
                "required": [
                  "email"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "WebAuthn options generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "challenge": {
                      "type": "string",
                      "description": "WebAuthn challenge string",
                      "example": "dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk"
                    },
                    "rpId": {
                      "type": "string",
                      "description": "Relying Party ID",
                      "example": "localhost"
                    },
                    "timeout": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991,
                      "description": "Challenge timeout in ms",
                      "example": 60000
                    },
                    "allowCredentials": {
                      "description": "Allowed authenticator credentials",
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Credential ID (base64url)"
                          },
                          "type": {
                            "type": "string",
                            "const": "public-key",
                            "example": "public-key"
                          }
                        },
                        "required": [
                          "id",
                          "type"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "userVerification": {
                      "type": "string",
                      "enum": [
                        "preferred",
                        "required",
                        "discouraged"
                      ],
                      "description": "User verification requirement",
                      "example": "preferred"
                    }
                  },
                  "required": [
                    "challenge",
                    "rpId",
                    "timeout",
                    "userVerification"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Passkey is not available for this device",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cws_session",
        "description": "HMAC-signed session cookie issued by the web dashboard login flow. Obtained via POST /api/auth/login (Server Action) or Google OAuth callback."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Ed25519-signed JWT issued by the mobile authentication endpoints. Obtained via POST /api/mobile/v1/auth/password or /api/mobile/v1/auth/google."
      },
      "pendingSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cws_2fa_pending",
        "description": "Short-lived HMAC-signed pending session cookie set during MFA challenge. Valid for 5 minutes. Used by WebAuthn and TOTP verification endpoints."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "A human-readable error message",
            "example": "Resource not found"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    }
  }
}